+++ directmemory-cache/src/main/java/org/apache/directmemory/memory/allocator/FixedSizeByteBufferAllocator.java
@@ -181,7 +181,7 @@ public class FixedSizeByteBufferAllocator
     }
 
     @Override
-    public int getCapacity()
+    public long getCapacity()
     {
         return totalSize;
     }
